+++ src/utils/removeElementFromArray.ts
