+++ src/utils/truncateLineWithEllipses.ts
